Define the "unwrap" methods in the AmqpConnection, AmqpSender, and AmqpReceiver classes.
These methods give access to the underlying Proton objects. In some cases, it is required to retrieve information on the links (sender/receiver) and open/close them.
The AmqpConnection unwrap method existed already but was not exposed in the interface. This commit exposes it in the interface.

Because the returned objects are not "polyglot", it uses the @GenIgnore(PERMITTED_TYPE)

Signed-off-by: Clement Escoffier <clement@apache.org>